Move add command tests into add packages
