+++ src/test/java/seedu/ddd/logic/parser/add/AddClientCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.ddd.logic.parser;
+package seedu.ddd.logic.parser.add;
 
 import static seedu.ddd.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.ddd.logic.commands.CommandTestUtil.CLIENT_FLAG;
@@ -42,6 +42,8 @@ import org.junit.jupiter.api.Test;
 
 import seedu.ddd.logic.Messages;
 import seedu.ddd.logic.commands.AddContactCommand;
+import seedu.ddd.logic.parser.AddClientCommandParser;
+import seedu.ddd.logic.parser.CommandParserTestUtil;
 import seedu.ddd.model.common.Name;
 import seedu.ddd.model.common.Tag;
 import seedu.ddd.model.contact.client.Client;
