+++ src/test/java/seedu/ddd/logic/parser/add/AddEventCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.ddd.logic.parser;
+package seedu.ddd.logic.parser.add;
 
 import static seedu.ddd.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.ddd.logic.commands.CommandTestUtil.CLIENT_FLAG;
@@ -45,6 +45,8 @@ import org.junit.jupiter.api.Test;
 
 import seedu.ddd.logic.Messages;
 import seedu.ddd.logic.commands.AddEventCommand;
+import seedu.ddd.logic.parser.AddEventCommandParser;
+import seedu.ddd.logic.parser.CommandParserTestUtil;
 import seedu.ddd.model.common.Id;
 import seedu.ddd.model.common.Name;
 import seedu.ddd.model.event.common.Date;
